mcp_server/projects/fix-teams-adaptive-cards/requirements.txt
msal>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.4.0
